MAINT: move rv_discrete and rv_continuous into the distribution infrastructure module and prune redundant imports
